refactor: package backend for render deployment
