+++ backend/Dockerfile
@@ -10,4 +10,4 @@ WORKDIR /app
 COPY --from=builder /usr/local /usr/local
 COPY . .
 ENV PATH=/usr/local/bin:$PATH
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "api.main:app", "--host", "0.0.0.0", "--port", "8000"]
