+++ backend/api/auth.py
@@ -2,14 +2,15 @@ import os
 import uuid
 from datetime import datetime, timedelta, timezone
 
-from fastapi import APIRouter, HTTPException
-from jose import jwt
+from fastapi import APIRouter, HTTPException, Request
+from jose import JWTError, jwt
 from passlib.context import CryptContext
 from pydantic import BaseModel, EmailStr
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from ..models import User
+from . import db
+from .models import User
 
 JWT_SECRET = os.getenv("JWT_SECRET", "change-me-please")
 JWT_ALG = "HS256"
@@ -34,10 +35,26 @@ def _create_token(uid: str) -> str:
     return jwt.encode({"sub": uid, "exp": exp}, JWT_SECRET, algorithm=JWT_ALG)
 
 
+def get_current_user(request: Request) -> User:
+    auth = request.headers.get("Authorization", "")
+    if not auth.startswith("Bearer "):
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    token = auth.split(" ", 1)[1]
+    try:
+        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALG])
+    except JWTError as exc:  # pragma: no cover - security check
+        raise HTTPException(status_code=401, detail="Invalid token") from exc
+    uid = payload.get("sub")
+    with Session(db.engine) as s:
+        user = s.get(User, uid)
+        if not user:
+            raise HTTPException(status_code=401, detail="User not found")
+        return user
+
+
 @router.post("/register", response_model=TokenResp)
 def register(body: AuthBody) -> TokenResp:
-    from .. import main as main_module
-    with Session(main_module.engine) as s, s.begin():
+    with Session(db.engine) as s, s.begin():
         if s.scalar(select(User).where(User.email == body.email)):
             raise HTTPException(status_code=400, detail="Email exists")
         user = User(
@@ -52,8 +69,7 @@ def register(body: AuthBody) -> TokenResp:
 
 @router.post("/login", response_model=TokenResp)
 def login(body: AuthBody) -> TokenResp:
-    from .. import main as main_module
-    with Session(main_module.engine) as s:
+    with Session(db.engine) as s:
         user = s.scalar(select(User).where(User.email == body.email))
         if not user or not pwd_context.verify(body.password, user.password_hash):
             raise HTTPException(status_code=401, detail="Invalid credentials")
