+++ backend/api/db.py
@@ -1,6 +1,7 @@
 import os
 from typing import Any
 from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
 from .models import Base
@@ -12,6 +13,7 @@ if DATABASE_URL.startswith("sqlite") and ":memory:" in DATABASE_URL:
 if os.getenv("SERVERLESS_DB") == "true":
     _engine_args.update(pool_size=3, max_overflow=3, pool_recycle=900)
 engine = create_engine(DATABASE_URL, **_engine_args)
+SessionLocal = sessionmaker(bind=engine, autocommit=False, autoflush=False, future=True)
 
 try:
     Base.metadata.create_all(engine)
