+++ run.bat
@@ -2,4 +2,4 @@
 cd /d %~dp0
 call .\backend\.venv\Scripts\activate
 python -m pip install -r .\backend\requirements.txt
-python -m uvicorn backend.main:app --reload
+python -m uvicorn api.main:app --reload
